requirements/base.txt
# Base requirements for FuelLens
Flask==3.0.0
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.3
Flask-WTF==1.2.1
Flask-Mail==0.9.1
Flask-Migrate==4.0.5
Flask-Limiter==3.5.0
Flask-Cors==4.0.0
Flask-Caching==2.1.0
Flask-Session==0.5.0
SQLAlchemy==2.0.23
psycopg2-binary==2.9.7
bcrypt==4.0.1
PyJWT==2.8.0
python-dotenv==1.0.0
requests==2.31.0
Pillow==10.1.0
opencv-python==4.8.1.78
pytesseract==0.3.10
qrcode==7.4.2
pyzbar==0.1.9
geopy==2.4.1
folium==0.15.1
APScheduler==3.10.4
redis==5.0.1
celery==5.3.4
gunicorn==21.2.0